OS X: compile misc
This is a rollup of commits authored by Adar, JD, and me. These changes allow
more parts of the codebase to be compiled on OS X.

Reviewed-on: http://gerrit.cloudera.org:8080/1223
Tested-by: Internal Jenkins

+++ src/kudu/cfile/binary_dict_block.cc
@@ -150,7 +150,7 @@ Status BinaryDictBlockBuilder::AppendExtraInfo(CFileWriter* c_writer, CFileFoote
   return Status::OK();
 }
 
-uint64_t BinaryDictBlockBuilder::Count() const {
+size_t BinaryDictBlockBuilder::Count() const {
   return data_builder_->Count();
 }
